requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
oauthlib==3.2.2
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==3.8.0
urllib3==1.26.13
